refactor: group candidate boundaries
