+++ ARCHITECTURE.md
@@ -32,8 +32,9 @@ decision (and usually a demolition pass) instead of silent sprawl.
 | `evaluation/evidence.py` | 120 | evaluator-output validation and conversion into canonical trial results |
 | `evaluation/identity.py` | 150 | canonical task-set identity plus checkout and frozen-baseline comparability |
 | `evaluation/execution.py` | 220 | clean-checkout canonical evaluation: tree assertion, lifecycle, artifact loading, and score parsing |
-| `candidate_snapshot.py` | 80 | exact candidate Git tree construction, temporary materialization, and reviewed-tree commit verification |
-| `candidate_smoke.py` | 130 | run evaluator smoke against an exact candidate snapshot and persist redacted diagnostics |
+| `candidate/__init__.py` | 10 | candidate-boundary package marker |
+| `candidate/snapshot.py` | 80 | exact candidate Git tree construction, temporary materialization, and reviewed-tree commit verification |
+| `candidate/smoke.py` | 130 | run evaluator smoke against an exact candidate snapshot and persist redacted diagnostics |
 | `git.py` | 150 | thin git subprocess helpers — nothing evolve-specific |
 | `surface.py` | 150 | mutable-surface pattern matching and violation checks |
 | `report.py` | 182 | status/report rendering, best-ever recomputation, claim checklist |
@@ -55,7 +56,7 @@ each workspace, immutable there because it sits outside the mutable surface
 | `frozen/sdk.py` | 420 | Python operator entrypoint and file-contract IO; no library algorithm policy |
 | `frozen/meta_eval.py` | 130 | self-modification admission gate: confound-free replay of old vs new operator surface (mechanism 1) |
 
-Total `src/evolve/` budget: **5320 lines**. The increase admits the explicit evaluation-package
+Total `src/evolve/` budget: **5330 lines**. The increase admits the explicit evaluation-package
 boundaries and the redacted trace-analysis boundary between rollout and feedback assembly; if the mechanism wants to
 grow past that, something belongs in a workspace operator instead —
 that is the spec's rule, not a style preference.
