+++ src/evolve/candidate/__init__.py
@@ -1,0 +1,1 @@
+"""Candidate snapshot and smoke boundaries."""
